Created new file for namespacing_helpers, removed argument in routes template
