+++ lib/generators/administrate/routes/routes_generator.rb
@@ -6,13 +6,14 @@ end
 
 require "rails/generators/base"
 require "administrate/generator_helpers"
+require "administrate/namespacing_helpers"
 require "administrate/namespace"
 
 module Administrate
   module Generators
     class RoutesGenerator < Rails::Generators::Base
       include Administrate::GeneratorHelpers
-      DEFAULT_INDENT = " " * 2
+      include Administrate::NamespacingHelpers
       INVALID_DATABASE_MODELS_LIST = [
         "ActiveRecord::SchemaMigration",
         "ActiveRecord::InternalMetadata",
@@ -53,37 +54,6 @@ module Administrate
         end
       end
 
-      def dashboard_resources_hash
-        output = [{}, []]
-        dashboard_resources.each do |resource_path|
-          split_path = resource_path.split("/")
-          output = [
-            add_namespace_to_hash(split_path, output[0]),
-            add_resources_to_hash(split_path, output[1]),
-          ]
-        end
-        output
-      end
-
-      def add_namespace_to_hash(split_path, current_state)
-        if split_path.length > 1
-          current_state[split_path[0]] ||= [{}, []]
-          current_namespace = current_state[split_path[0]]
-          current_namespace = [
-            add_namespace_to_hash(split_path[1..], current_namespace[0]),
-            add_resources_to_hash(split_path[1..], current_namespace[1]),
-          ]
-        end
-        current_state
-      end
-
-      def add_resources_to_hash(split_path, current_state)
-        if split_path.length == 1
-          current_state.push(split_path[0])
-        end
-        current_state
-      end
-
       def valid_dashboard_models
         database_models - (invalid_dashboard_models + excluded_models).uniq
       end
@@ -108,35 +78,6 @@ module Administrate
 
       def unnamed_constants
         ActiveRecord::Base.descendants.reject { |d| d.name == d.to_s }
-      end
-
-      def generate_resource_routes(namespace)
-        output_string = ""
-        namespace[1].each do |resource|
-          output_string += "#{DEFAULT_INDENT}resources :#{resource}\n"
-        end
-        output_string += "\n#{generate_namespace_routes(namespace[0])}"
-        output_string
-      end
-
-      def generate_namespace_routes(hash, indent = 1)
-        output_string = ""
-        indentation = DEFAULT_INDENT * indent
-        hash.each do |namespace, namespace_resource|
-          nested_namespaces = namespace_resource[0]
-          nested_resources = namespace_resource[1]
-          output_string += "#{indentation}namespace :#{namespace} do\n"
-          nested_resources.each do |resource|
-            resource_string = "#{DEFAULT_INDENT}resources :#{resource}"
-            output_string += "#{indentation}#{resource_string}\n"
-          end
-          if !nested_namespaces.empty?
-            namespaces_string = generate_namespace_routes(nested_namespaces, indent + 1)
-            output_string += "\n#{namespaces_string}\n"
-          end
-          output_string += "#{indentation}end\n"
-        end
-        output_string
       end
 
       def dashboard_routes
