+++ lib/generators/administrate/routes/templates/routes.rb.erb
@@ -1,4 +1,4 @@
 namespace :<%= namespace %> do
-<%= generate_resource_routes(dashboard_resources_hash) %>
+<%= generate_resource_routes() %>
   root to: "<%= dashboard_resources.first %>#index"
 end
